requirements.txt
qwen-tts>=0.1.0
torch>=2.1.0
soundfile>=0.12.0
websockets>=12.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
pydantic>=2.0.0
numpy>=1.24.0
pydub>=0.25.0
